Fixed an obscure error where you're not allowed to overload the toString(), even if the overload arguments are optional, by moving the behaviour into another function altogether. Other than that, fixed up the 'hello' example so that weird characters weren't showing up for an ASCII space.

+++ src/nel/parse/parser.d
@@ -161,7 +161,7 @@ class Parser
                 else
                 {
                     // Pop previous scanner off stack.
-                    ulong top = includeStack.length - 1;
+                    uint top = includeStack.length - 1;
                     scanner = includeStack[top];
                     includeStack.length = top;
                     // Ready a new token for the scanner.
